fix(client): rework 3rd party libraries initialization
