--- client/public/css/main.css
@@ -1,4 +1,0 @@
-.vuln-list {
-    text-align: left;
-    padding: 50px;
-}
