+++ client/public/index.html
@@ -124,9 +124,6 @@
     <link href="assets/vendor/venobox/venobox.css" rel="stylesheet" />
     <link href="assets/vendor/aos/aos.css" rel="stylesheet" />
 
-    <!-- Template Main CSS File -->
-    <link href="assets/css/style.css" rel="stylesheet" />
-
     <!-- Fontfaces CSS-->
     <link
       href="vendor/font-awesome-4.7/css/font-awesome.min.css"
@@ -154,6 +151,7 @@
     <link href="vendor/slick/slick.css" rel="stylesheet" media="all" />
 
     <!-- Main CSS-->
+    <link href="assets/css/style.css" rel="stylesheet" media="all" />
     <link href="css/theme.css" rel="stylesheet" media="all" />
   </head>
   <body>
@@ -186,8 +184,5 @@
     <script src="vendor/animsition/animsition.min.js"></script>
     <script src="vendor/counter-up/jquery.waypoints.min.js"></script>
     <script src="vendor/counter-up/jquery.counterup.min.js"></script>
-
-    <!-- Main JS-->
-    <script src="js/main.js"></script>
   </body>
 </html>
