+++ client/src/index.tsx
@@ -1,13 +1,26 @@
-import React from 'react';
+import React, { useEffect } from 'react';
 import { createRoot } from 'react-dom/client';
 import { BrowserRouter } from 'react-router-dom';
 import { AppRoutes } from './router/AppRoutes';
+import { initDependencies } from './main';
+
+// New wrapper component
+function AppWrapper({ onRender }: { onRender: () => void }) {
+  useEffect(() => {
+    onRender();
+  }, [onRender]);
+
+  return (
+    <BrowserRouter>
+      <AppRoutes />
+    </BrowserRouter>
+  );
+}
 
 // eslint-disable-next-line @typescript-eslint/no-non-null-assertion
-createRoot(document.getElementById('root')!).render(
+const root = createRoot(document.getElementById('root')!);
+root.render(
   // <React.StrictMode>
-  <BrowserRouter>
-    <AppRoutes />
-  </BrowserRouter>
+  <AppWrapper onRender={() => initDependencies()} />
   // </React.StrictMode>
 );
