+++ client/src/interfaces/Product.ts
@@ -1,4 +1,5 @@
 export interface Product {
+  id: number;
   name: string;
   category: string;
   photoUrl: string;
